Correct a possible mistake in last commit
Before submitting my completed Q1, I renamed the file to indicate that it was completed. I now see that the pre-work in my fork of the dsp links to the old file name/location, so I an renaming the file again to the old name (2-4-cohens_d.md). This is also a way for me to use git more and get more comfortable with it.